# Fix bluetooth_proxy heap allocations during BLE scanning