--- a/esphome/components/api/api_options.proto
+++ b/esphome/components/api/api_options.proto
@@ -26,4 +26,5 @@ extend google.protobuf.MessageOptions {
 
 extend google.protobuf.FieldOptions {
     optional string field_ifdef = 1042;
+    optional uint32 fixed_array_size = 50007;
 }