--- a/esphome/components/api/api_pb2_dump.cpp
+++ b/esphome/components/api/api_pb2_dump.cpp
@@ -3132,7 +3132,7 @@ void BluetoothLERawAdvertisement::dump_to(std::string &out) const {
   out.append("\n");
 
   out.append("  data: ");
-  out.append(format_hex_pretty(reinterpret_cast<const char *>(this->data), this->data_len));
+  out.append(format_hex_pretty(this->data, this->data_len));
   out.append("\n");
   out.append("}");
 }