# What does this implement/fix?

This PR eliminates the remaining heap allocations in the bluetooth_proxy component during BLE advertisement scanning. Previously, the component was allocating memory ~160 times per second during heavy BLE scanning, causing memory fragmentation and performance degradation.

The optimization introduces:
- Fixed-size arrays for BLE advertisement data instead of std::string
- Pre-allocated advertisement objects that are reused (with 50% initial reservation)
- Zero-allocation operation in steady state
- 81% reduction in BLE advertisement processing time
- 42.3% reduction in total component runtime
- Memory usage remains roughly the same

## Types of changes

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):**

- Continuation of memory optimization efforts from previous bluetooth_proxy PRs
- Addresses heap churn during heavy BLE scanning scenarios

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):**

- N/A (internal optimization, no user-facing changes)

## Test Environment

- [x] ESP32
- [x] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Example entry for `config.yaml`:

```yaml
# No configuration changes required - this is an internal optimization
bluetooth_proxy:
  active: true
```

## Performance Results

### Before optimization:
```
esp32_ble_tracker: count=6598, avg=0.09ms, max=6ms, total=586ms
Free Memory: 149,812 bytes
```

### After optimization:
```
esp32_ble_tracker: count=6686, avg=0.02ms, max=4ms, total=110ms  
Free Memory: ~150,000 bytes (roughly the same as before)
```
*Note: During BLE traffic spikes, may use up to ~1.3KB more RAM to maintain pre-allocated objects and avoid allocation churn*

- **78% faster** average processing time (0.09ms → 0.02ms)
- **81% less total time** spent processing advertisements
- **42.3% total runtime decrease** across all components (1,557ms → 898ms per 60s)
- Zero heap allocations during steady-state operation
- Fixed memory footprint (no fragmentation)

## Technical Details

1. **API Protocol Buffer Changes**:
   - Added `fixed_array_size` option to protobuf field options
   - Changed `BluetoothLERawAdvertisement.data` from `bytes` to fixed array `uint8_t data[62]`

2. **Memory Management**:
   - Pre-allocate response object with 50% reserved capacity (8 slots, grows to 16 as needed)
   - Reuse advertisement objects by overwriting data
   - Pool for overflow scenarios (grows only if needed - no pre-allocation)
   - Total maximum of 16 advertisements at any time (enforced by FLUSH_BATCH_SIZE)
   - No allocations after initial setup
   - Memory neutral in typical use - same baseline memory usage
   - During BLE traffic spikes: slightly more RAM used but objects are retained to prevent allocation churn
   - No memory fragmentation - predictable memory footprint

3. **Code Generation Updates**:
   - Added `FixedArrayBytesType` to handle fixed-size byte arrays
   - Generate `memcpy` instead of string operations
   - Fixed size calculation to match encode behavior

## Checklist:
  - [x] The code change is tested and works locally.
  - [x] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).