@bdraco 2025-07-18
still works
```
[14:17:48][VV][api.service:013]: send_message bluetooth_le_raw_advertisements_response: BluetoothLERawAdvertisementsResponse {
[14:17:48][VV][api.service:013]:   advertisements: BluetoothLERawAdvertisement {
[14:17:48][VV][api.service:013]:   address: 244579075710878
[14:17:48][VV][api.service:013]:   rssi: -57
[14:17:48][VV][api.service:013]:   address_type: 1
[14:17:48][VV][api.service:013]:   data: 02.01.06.0D.FF.69.09.DE.71.80.ED.5F.9E.31.40.08.20 (17)
[14:17:48]}
[14:17:48]  advertisements: BluetoothLERawAdvertisement {
[14:17:48]  address: 57911552599661
[14:17:48]  rssi: -81
[14:17:48]  address_type: 0
[14:17:48]  data: 02.01.06.03.03.12.18.10.09.4C.4F.4F.4B.69.6E.5F.39.38.46.33.33.31.36.33.03.19.C1.03 (28)
[14:17:48]}

```


but now I have to come up with a fix for the scheduler that doesn't use atomics for libretiny 😢