@bdraco 2025-07-19
From a production device. Before and after. You can see the reduction in heap churn in the memory graph. While it wasn't that much before because we fixed the churn in the BLE queue, its near none now. 

<img width="576" height="327" alt="Screenshot 2025-07-18 at 7 34 14 PM" src="https://github.com/user-attachments/assets/5005415b-4b38-460d-a877-51df43f513cd" />